change project tree structure for pages
